Move data prep for segments to features helper file
